requirements.txt
-i https://pypi.org/simple
attrs==23.2.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
clickclick==20.10.2
connexion[swagger-ui]==2.14.1; python_version >= '3.6'
flask==2.2.2; python_version >= '3.7'
flask-cors==4.0.0
flask-httpauth==3.2.3
flask-marshmallow[sqlalchemy]==0.14.0
flask-sqlalchemy==3.0.3
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.6; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.0; python_version >= '3.8'
marshmallow-sqlalchemy==1.0.0
packaging==23.2; python_version >= '3.7'
passlib==1.7.1
psycopg2-binary==2.9.9; python_version >= '3.7'
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.33.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.18.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.27; python_version >= '3.7'
swagger-ui-bundle==0.0.9
typing-extensions==4.10.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
werkzeug==2.3.8; python_version >= '3.8'
